+++ apps/admin-x-activitypub/src/views/Feed/Feed.tsx
@@ -1,33 +1,9 @@
-import APAvatar from '@components/global/APAvatar';
-import FeedItem from '@components/feed/FeedItem';
-import Layout from '@components/layout';
-import NewPostModal from '@views/Feed/components/NewPostModal';
-import NiceModal from '@ebay/nice-modal-react';
+import FeedList from './components/FeedList';
 import React, {useEffect, useRef} from 'react';
-import Separator from '@components/global/Separator';
-import {ActorProperties} from '@tryghost/admin-x-framework/api/activitypub';
-import {Button, LucideIcon} from '@tryghost/shade';
-import {EmptyViewIcon, EmptyViewIndicator} from '@src/components/global/EmptyViewIndicator';
-import {LoadingIndicator} from '@tryghost/admin-x-design-system';
-import {isPendingActivity} from '@utils/pending-activity';
 import {
     useFeedForUser,
     useUserDataForUser
 } from '@hooks/use-activity-pub-queries';
-import {useNavigate} from '@tryghost/admin-x-framework';
-
-const FeedInput: React.FC<{user?: ActorProperties}> = ({user}) => {
-    return (
-        <>
-            <div className='relative my-5 w-full'>
-                <div className='pointer-events-none absolute left-4 top-4'>
-                    <APAvatar author={user as ActorProperties} />
-                </div>
-                <Button aria-label='New post' className='text inset-0 h-[72px] w-full justify-start rounded-lg bg-white pl-[68px] text-left text-[1.5rem] font-normal tracking-normal text-gray-500 shadow-[0_5px_24px_0px_rgba(0,0,0,0.02),0px_2px_5px_0px_rgba(0,0,0,0.07),0px_0px_1px_0px_rgba(0,0,0,0.25)] transition-all hover:bg-white hover:shadow-[0_5px_24px_0px_rgba(0,0,0,0.05),0px_14px_12px_-9px_rgba(0,0,0,0.07),0px_0px_1px_0px_rgba(0,0,0,0.25)] dark:border dark:border-gray-925 dark:bg-black dark:shadow-none dark:hover:border-gray-800 dark:hover:bg-black dark:hover:shadow-none' onClick={() => NiceModal.show(NewPostModal)}>What&apos;s new?</Button>
-            </div>
-        </>
-    );
-};
 
 const Feed: React.FC = () => {
     const {feedQuery} = useFeedForUser({enabled: true});
@@ -72,82 +48,15 @@ const Feed: React.FC = () => {
     // Calculate the index at which to place the loadMoreRef - This will place it ~75% through the list
     const loadMoreIndex = Math.max(0, Math.floor(activities.length * 0.75) - 1);
 
-    const navigate = useNavigate();
-
-    return (
-        <Layout>
-            <div className='flex w-full flex-col'>
-                <div className='w-full'>
-                    {activities.length > 0 ? (
-                        <div className='my-4'>
-                            <div className='mx-auto flex items-start gap-11'>
-                                <div className='flex w-full min-w-0 flex-col items-center'>
-                                    <div className='flex w-full min-w-0 max-w-[620px] flex-col items-start'>
-                                        <FeedInput user={user} />
-                                        <ul className='mx-auto flex w-full flex-col px-4'>
-                                            {activities.map((activity, index) => (
-                                                <li
-                                                    // eslint-disable-next-line react/no-array-index-key
-                                                    key={`${activity.id}-${activity.type}-${index}`} // We are using index here as activity.id is cannot be guaranteed to be unique at the moment
-                                                    data-test-view-article
-                                                >
-                                                    <FeedItem
-                                                        actor={activity.actor}
-                                                        allowDelete={activity.object.authored}
-                                                        commentCount={activity.object.replyCount ?? 0}
-                                                        isLoading={isLoading}
-                                                        isPending={isPendingActivity(activity.id)}
-                                                        layout={'feed'}
-                                                        object={activity.object}
-                                                        repostCount={activity.object.repostCount ?? 0}
-                                                        type={activity.type}
-                                                        onClick={() => {
-                                                            navigate(`/feed-rr/${encodeURIComponent(activity.id)}`);
-                                                        }}
-                                                        onCommentClick={() => {
-                                                            navigate(`/feed-rr/${encodeURIComponent(activity.id)}?focusReply=true`);
-                                                        }}
-                                                    />
-                                                    {index < activities.length - 1 && (
-                                                        <Separator />
-                                                    )}
-                                                    {index === loadMoreIndex && (
-                                                        <div ref={loadMoreRef} className='h-1'></div>
-                                                    )}
-                                                </li>
-                                            ))}
-                                            {isFetchingNextPage && (
-                                                <li className='flex flex-col items-center justify-center space-y-4 text-center'>
-                                                    <LoadingIndicator size='md' />
-                                                </li>
-                                            )}
-                                        </ul>
-                                        <div ref={endLoadMoreRef} className='h-1'></div>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                    ) : (
-                        <div className='flex w-full flex-col items-center gap-10'>
-                            <div className='mt-4 flex w-full max-w-[620px] flex-col items-center'>
-                                <FeedInput user={user} />
-                                <div className='mt-[-128px]'>
-                                    <EmptyViewIndicator>
-                                        <EmptyViewIcon><LucideIcon.Hash /></EmptyViewIcon>
-                                        <div>The Feed is the stream of thoughts and <span className='text-black dark:text-white'>bite-sized updates</span> from people you follow in the Social Web. It&apos;s looking a little empty right now but once the people you follow start posting, their updates will show up here.</div>
-                                        <Button className='text-white dark:text-black' onClick={() => NiceModal.show(NewPostModal)}>
-                                            <LucideIcon.FilePen />
-                                                Write your first note
-                                        </Button>
-                                    </EmptyViewIndicator>
-                                </div>
-                            </div>
-                        </div>
-                    )}
-                </div>
-            </div>
-        </Layout>
-    );
+    return <FeedList
+        activities={activities}
+        endLoadMoreRef={endLoadMoreRef}
+        isFetchingNextPage={isFetchingNextPage}
+        isLoading={isLoading}
+        loadMoreIndex={loadMoreIndex}
+        loadMoreRef={loadMoreRef}
+        user={user!}
+    />;
 };
 
 export default Feed;
